move script and add yarn command
