+++ packages/browser/package.json
@@ -76,6 +76,7 @@
     "test:unit:watch": "karma start test/unit/karma.conf.js --auto-watch --no-single-run",
     "test:integration": "test/integration/run.js",
     "test:integration:watch": "test/integration/run.js --watch",
+    "test:integration:checkbrowsers": "node scripts/checkbrowsers.js",
     "test:manual": "node test/manual/npm-build.js && rm test/manual/tmp.js",
     "size:check": "run-p size:check:es5 size:check:es6",
     "size:check:es5": "cat build/bundle.min.js | gzip -9 | wc -c | awk '{$1=$1/1024; print \"ES5: \",$1,\"kB\";}'",
